+++ unity-samples/Assets/Scripts/SetCustomAttributesArray.cs
@@ -15,6 +15,8 @@ public class SetCustomAttributesArray : MonoBehaviour {
   public InputField AddStringToArrayStringField;
   public InputField RemoveStringFromArrayKeyField;
   public InputField RemoveStringFromArrayStringField;
+  public InputField NestedArrayKeyField;
+  
 
   public void OnBackButtonClick() {
     SceneManager.LoadScene(Constants.CustomAttributesSettingScene);
@@ -43,6 +45,29 @@ public class SetCustomAttributesArray : MonoBehaviour {
       Debug.Log(String.Format("Removing string {0} from array with key {1}", 
       RemoveStringFromArrayStringField.text, RemoveStringFromArrayKeyField.text));
       Appboy.AppboyBinding.RemoveFromCustomUserAttributeArray(RemoveStringFromArrayKeyField.text, RemoveStringFromArrayStringField.text);
+    }
+  }
+
+  public void OnCreateNestedArrayButtonClick() {
+    if (UtilityMethods.textIsValid(NestedArrayKeyField.text)) {
+      Dictionary<string, object> object1 = new Dictionary<string, object>();
+      object1.Add("key int", 2);
+      object1.Add("key string", "first value");
+      object1.Add("key bool", true);
+      object1.Add("key float", 20.2);
+
+      Dictionary<string, object> object2 = new Dictionary<string, object>();
+      object2.Add("key int", 3);
+      object2.Add("key string", "second value");
+      object2.Add("key bool", false);
+      object2.Add("key float", 23.2);
+
+      List<Dictionary<String, object>> listOfStuff = new List<Dictionary<String, object>>();
+      listOfStuff.Add(object1);
+      listOfStuff.Add(object2);
+
+      Debug.Log(String.Format("Creating Nested Array {0}", NestedArrayKeyField.text));
+      Appboy.AppboyBinding.SetCustomUserAttribute(NestedArrayKeyField.text, listOfStuff);
     }
   }
 }